fix : permissions issuses and add more permissions

+++ app/Http/Controllers/Admin/ClientController.php
@@ -2,9 +2,11 @@
 
 namespace App\Http\Controllers\Admin;
 
+use App\Enums\PermissionEnum;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreClientRequest;
 use App\Http\Requests\UpdateClientRequest;
+use App\Models\Client;
 use App\Repositories\SalesOfficerRepo;
 use App\Services\CountService;
 use Illuminate\Http\Request;
@@ -28,6 +30,8 @@ class ClientController extends Controller
     }
     public function index(Request $request)
     {
+        $this->authorize(PermissionEnum::CLIENT(), [Client::class]);
+
         $salesOfficers = $this->salesOfficerRepo->getAllSalesOfficers();
         if ($request->has('query')) {
             $searcData = $request->all();
@@ -41,6 +45,8 @@ class ClientController extends Controller
     }
     public function create()
     {
+        $this->authorize(PermissionEnum::CLIENT_CREATE(), [Client::class]);
+
         $salesOfficers = $this->clientRepository->getSalesOfficers();
         return view("admin.client.create", compact("salesOfficers"));
     }
@@ -51,11 +57,15 @@ class ClientController extends Controller
     }
     public function show($id)
     {
+        $this->authorize(PermissionEnum::CLIENT_VIEW(), [Client::class]);
+
         $data = $this->clientRepository->show($id);
         return view('admin.client.show', compact('data'));
     }
     public function edit($id)
     {
+        $this->authorize(PermissionEnum::CLIENT_EDIT(), [Client::class]);
+
         $data = $this->clientRepository->show($id);
         // dd($data);
         return view('admin.client.edit', compact('data'));
@@ -67,11 +77,15 @@ class ClientController extends Controller
     }
     public function delete($id)
     {
+        $this->authorize(PermissionEnum::CLIENT_DELETE(), [Client::class]);
+
         $this->clientRepository->delete($id);
         return redirect('admin/client')->with('success', 'Record Deleted Successfully.');
     }
     public function getInstallments($id)
     {
+        $this->authorize(PermissionEnum::CLIENT_INSTALLMENT_VIEW(), [Client::class]);
+
         $data = $this->clientRepository->getCashInstallments($id);
         $chequeInstallments = $data[1];
         $cashInstallments = $data[0];
@@ -79,12 +93,16 @@ class ClientController extends Controller
     }
     public function installmentUpdate($id)
     {
+        $this->authorize(PermissionEnum::CLIENT_INSTALLMENT_STATUS(), [Client::class]);
+
         $data = $this->clientRepository->updateInstallmentStatus($id);
         return redirect()->back()->with('success', 'Status Update Successfully.');
     }
 
     public function addNewCashInstallment(Request $data, $id)
     {
+        $this->authorize(PermissionEnum::CLIENT_CASH_INSTALLMENT_ADD(), [Client::class]);
+
         $customCashInstallment = $this->plotInstallmentRepository->addCustomCashInstallment($data, $id);
         if ($customCashInstallment == false) {
             return redirect()->back()->with('error', 'Installment Amount Is More Than Total Amount.');
@@ -94,6 +112,8 @@ class ClientController extends Controller
     }
     public function addNewChequeInstallment(Request $data, $id)
     {
+        $this->authorize(PermissionEnum::CLIENT_CHECK_INSTALLMENT_ADD(), [Client::class]);
+
         $customChequeInstallment = $this->plotInstallmentRepository->addCustomChequeInstallment($data, $id);
         if ($customChequeInstallment == false) {
             return redirect()->back()->with('error', 'Installment Amount Is More Than Total Amount.');
