+++ app/Http/Controllers/Admin/OfficeEmployeeController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\Admin;
 
+use App\Enums\PermissionEnum;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\AddOfficeEmployeeRequest;
 use App\Models\AdminOfficeEMployee;
@@ -17,11 +18,15 @@ class OfficeEmployeeController extends Controller
     }
     public function index()
     {
+        $this->authorize(PermissionEnum::EMPLOYEE(), [AdminOfficeEMployee::class]);
+
         $data = $this->employeeRepository->all();
         return view("admin.officeEmployee.index" , compact("data"));
     }
     public function create()
     {
+        $this->authorize(PermissionEnum::EMPLOYEE_CREATE(), [AdminOfficeEMployee::class]);
+
         return view("admin.officeEmployee.create");
     }
     public function store(AddOfficeEmployeeRequest $request)
@@ -32,11 +37,15 @@ class OfficeEmployeeController extends Controller
 
     public function show($id)
     {
+        $this->authorize(PermissionEnum::EMPLOYEE_VIEW(), [AdminOfficeEMployee::class]);
+
         $data = $this->employeeRepository->find($id);
         return view("admin.officeEmployee.show", data: compact(var_name: 'data'));
     }
     public function edit($id)
     {
+        $this->authorize(PermissionEnum::EMPLOYEE_EDIT(), [AdminOfficeEMployee::class]);
+
         $data = $this->employeeRepository->find($id);
         return view("admin.officeEmployee.edit", data: compact(var_name: 'data'));
     }
@@ -48,6 +57,8 @@ class OfficeEmployeeController extends Controller
 
     public function delete($id)
     {
+        $this->authorize(PermissionEnum::EMPLOYEE_DELETE(), [AdminOfficeEMployee::class]);
+
         $employee = AdminOfficeEMployee::find($id);
         $employee->delete();
         return redirect()->back()->with('success', 'Record Deleted Successfully.');
